Add stc, clc, cmc, sahf, fnstsw
Added a couple of trivial instructions that manipulate the carry flag. Modified adc and add to store their results in a temp register and then compare its result against the initial value of operand(0) to obtain the value of cf, which helps with analyzing code that uses arbitrary precision arithmetic. Also added sahf, but not lahf yet because I haven't researched what the values of the unspecified flags are after its execution